this is all just dump
This is an example of a dump file. Just commit everything. this is created bc if you just load into the unity package, its gonna dump files in here just because they were touched if that makes sense. so always commit after launching. this is  how people forget to commit after launching and it gets corrupted, or people forgetting to pull from the origin to update the unity package.